# docs/reference/api-spec.yaml
openapi: 3.0.3
info:
  title: lole Restaurant OS API
  version: 1.0.0
  description: |
    Enterprise-grade Restaurant Operating System API for Ethiopian restaurants.
    Supports offline-first operations with synchronization capabilities.
  contact:
    name: lole Engineering
    email: api@lole.restaurant
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.lole.app
    description: Production server (Global)
  - url: https://api.staging.lole.app
    description: Staging server (Testing)
  - url: http://localhost:3000/api
    description: Local development server
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
...
paths:
...
  /api/v1/merchant/operations/kds/orders/{orderId}:
    patch:
      tags: [KDS]
      summary: Update KDS order
      operationId: updateKDSOrder
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: KDS order updated
...